Changed problems when testing the website in browser.
Added space to make the subtitle more attention grabbing. Put the layout style in index page,because when testing on safari the style only worked once I am on that page. meaning the other pages were lagging a bit, so changed it to index as it would prevent the lagging and the layout would appear quickly.