Improved website.
By adding text styles, colours and hover effect on nav to improve user experience. The third colour is too light to be seen by users so I chose to not use it on the content. Instead used it on the hover with a dark background so it can be seen.

Tested on safari and chrome to see which font best suit the nav.