Testing on safari and chrome to see the output of different codes.
Tried codes to get my wanted outcome. Changed hr style to look better and clear.

Styled the page to improve its aesthetics to be more appealing to the end user.

Due to user feedback, I change nav to be at the top of the header and spaced out the sections. The end user likes the layout, so I will try to use it as they are the ones using the website most.