Improve the website by adding styles.
I first went with the user feedback on the colours and tested them in safari and chrome, but they were too light, so I decided to use similar colours as the end user liked them but a bit darker as the client said they are too light.

Imported the fonts from google fonts.

Due to feedback, I tried the fonts first and tested it in safari. It was hard to read, so I will follow the feedback I got and choose different fonts completely.