Change the layout as there is not much text from the client.
Tested in safari and chorme to see if it looks ok and it did. Also reduced the content in this style by using only two columns, which was a feedback I got to improve my design.

Tried out the fonts first, and they don't match just like the feedback I got. So I will change the fonts to a different font.