Improved the website by adding colour.
I tried the colour combination that the end user says is good, but then when I tested on safari I agree with the stakeholder feedback that its a bit strong. So I decided to use different colours.

When I test the outcome, I saw the nav was still on top of the header and not under it, so I put the nav component into the main tag, and it worked when I tested again in safari.