Improved the website
By fixing some spelling errors, put the two groups names so students know which page they are on.

As well as improving the style by giving them font family for the text to look better.

Added some spacing so the subtitle can be separate from the header, this improves the layout and look of the website.

Improved the code by using em instead of px for the sizes of fonts and spacing so its easier to compare different sizes.