improved the website by using hirachy
One page's text size hierachy is wrong, so changed it to be largerst font to smallerst. Improved it as it now  fits the convention of hierachy, and is more familiar to users.

By using this convention, it improves the usability and accessibility of the website as most users can understand it as its most familiar to them.