Tested in safari
The visual communication page text were larger than other pages, which is   a problem as its not consistent font size, so to change this and improve the consistence of the website, I found the problem and fixed it.
The problem was the main font size written before, so deleted that and fixed the font size problem.