Tested in safari and chrome and found some problems, then fixed it and improved the website more.
Change the position of the subtitles so they appear in the same place. Enlarged the font sizes to make them more attracting. Also tested mobile view, which looks good for users to use.

Tested on safari and chrome to see that when the screen decreased to a certain size only the home in the nav bar would appear. Which worked in both browser, and that ensured the function of the website is working.

Tested the link function, and they all worked well.

Improved the nav hover effect by changing its colour to be not as strong. Before the hover on the home was a different colour from the rest, by adding nav infront of the a made sure that the home hover effect was the same as other nav items. Also improved the function of the nav items by adding a transition so its slowly changing when the mouse hovers above than before where it changed too fast and didn't look good.