Stakeholder feedback said having check box could improve the usability of the website.
Tested the website with the checkbox on safari and found out that it was too small and not a nice look with the website. I decided to deleted it as aesthetic of the website would be more important for year 10 students as their attention are mainly on the appearances of things.